# data/binding: optimized convertion using strconv