--- a/data/binding/gen.go
+++ b/data/binding/gen.go
@@ -154,13 +154,9 @@ type stringFrom{{ .Name }} struct {
 //
 // Since: {{ .Since }}
 func {{ .Name }}ToString(v {{ .Name }}) String {
-{{- if .Format }}
-	return {{ .Name }}ToStringWithFormat(v, "{{ .Format }}")
-{{- else }}
 	str := &stringFrom{{ .Name }}{from: v}
 	v.AddListener(str)
 	return str
-{{- end }}
 }
 {{ if .Format }}
 // {{ .Name }}ToStringWithFormat creates a binding that connects a {{ .Name }} data item to a String and is
@@ -169,6 +165,10 @@ func {{ .Name }}ToString(v {{ .Name }}) String {
 //
 // Since: {{ .Since }}
 func {{ .Name }}ToStringWithFormat(v {{ .Name }}, format string) String {
+	if format == "{{ .Format }}" { // Same as not using custom formatting.
+		return {{ .Name }}ToString(v)
+	}
+	
 	str := &stringFrom{{ .Name }}{from: v, format: format}
 	v.AddListener(str)
 	return str
@@ -182,7 +182,11 @@ func (s *stringFrom{{ .Name }}) Get() (string, error) {
 {{ if .ToString }}
 	return {{ .ToString }}(val)
 {{- else }}
-	return fmt.Sprintf(s.format, val), nil
+	if s.format != "" {
+		return fmt.Sprintf(s.format, val), nil
+	}
+
+	return format{{ .Name }}(val), nil
 {{- end }}
 }
 
@@ -194,12 +198,20 @@ func (s *stringFrom{{ .Name }}) Set(str string) error {
 	}
 {{ else }}
 	var val {{ .Type }}
-	n, err := fmt.Sscanf(str, s.format+" ", &val) // " " denotes match to end of string
-	if err != nil {
-		return err
-	}
-	if n != 1 {
-		return errParseFailed
+	if s.format != "" {
+		n, err := fmt.Sscanf(str, s.format+" ", &val) // " " denotes match to end of string
+		if err != nil {
+			return err
+		}
+		if n != 1 {
+			return errParseFailed
+		}
+	} else {
+		new, err := parse{{ .Name }}(str)
+		if err != nil {
+			return err
+		}
+		val = new
 	}
 {{ end }}
 	old, err := s.from.Get()
@@ -239,13 +251,9 @@ type stringTo{{ .Name }} struct {
 //
 // Since: {{ .Since }}
 func StringTo{{ .Name }}(str String) {{ .Name }} {
-{{- if .Format }}
-	return StringTo{{ .Name }}WithFormat(str, "{{ .Format }}")
-{{- else }}
 	v := &stringTo{{ .Name }}{from: str}
 	str.AddListener(v)
 	return v
-{{- end }}
 }
 {{ if .Format }}
 // StringTo{{ .Name }}WithFormat creates a binding that connects a String data item to a {{ .Name }} and is
@@ -255,6 +263,10 @@ func StringTo{{ .Name }}(str String) {{ .Name }} {
 //
 // Since: {{ .Since }}
 func StringTo{{ .Name }}WithFormat(str String, format string) {{ .Name }} {
+	if format == "{{ .Format }}" { // Same as not using custom format.
+		return StringTo{{ .Name }}(str)
+	}
+	
 	v := &stringTo{{ .Name }}{from: str, format: format}
 	str.AddListener(v)
 	return v
@@ -269,12 +281,20 @@ func (s *stringTo{{ .Name }}) Get() ({{ .Type }}, error) {
 	return {{ .FromString }}(str)
 {{- else }}
 	var val {{ .Type }}
-	n, err := fmt.Sscanf(str, s.format+" ", &val) // " " denotes match to end of string
-	if err != nil {
-		return {{ .Default }}, err
-	}
-	if n != 1 {
-		return {{ .Default }}, errParseFailed
+	if s.format != "" {
+		n, err := fmt.Sscanf(str, s.format+" ", &val) // " " denotes match to end of string
+		if err != nil {
+			return {{ .Default }}, err
+		}
+		if n != 1 {
+			return {{ .Default }}, errParseFailed
+		}
+	} else {
+		new, err := parse{{ .Name }}(str)
+		if err != nil {
+			return {{ .Default }}, err
+		}
+		val = new
 	}
 
 	return val, nil
@@ -288,7 +308,12 @@ func (s *stringTo{{ .Name }}) Set(val {{ .Type }}) error {
 		return err
 	}
 {{- else }}
-	str := fmt.Sprintf(s.format, val)
+	var str string
+	if s.format != "" {
+		str = fmt.Sprintf(s.format, val)
+	} else {
+		str = format{{ .Name }}(val)
+	}
 {{ end }}
 	old, err := s.from.Get()
 	if str == old {