### Description:
<!-- A summary of the change included and which issue it addresses.
Please include any relevant motivation and background. -->

This updates the data bindings to use the `strconv` package instead of `fmt` for conversions, leading to improved performance (benchmarks to come). This is currently WIP and is just a proof of concept with generations scripts yet to have been updated to take this into effect.

Fixes #1735

### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [ ] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.