@andydotxyz 2021-04-07
Looks like a good change thanks.
I wonder (random thought) should we also use this version when someone manually specifies the default format?